==> 5 - Statistical tests.ipynb <==
from pathlib import Path

import numpy as np
import pandas as pd
import statsmodels.api as sm
import statsmodels.formula.api as smf
from scipy.stats import ttest_ind
from sklearn.preprocessing import StandardScaler
from statsmodels.iolib.summary2 import summary_col
from statsmodels.stats.multicomp import pairwise_tukeyhsd


data_dir = Path() / "results"
figs_dir = Path() / "figs"
figs_dir.mkdir(exist_ok=True)

# Statistics

cronbach_df = pd.read_csv(data_dir / "cronbach.csv")
cronbach_df_simple = cronbach_df.query("scale in ['MFQ', 'MFV']")

## Condition T-test

# claude ttest
ttest_ind(
    cronbach_df_simple.query("agent == 'Claude 2.1' and condition == 'qv'")["alpha"].values,
    cronbach_df_simple.query("agent == 'Claude 2.1' and condition == 'vq'")["alpha"].values,
)

# old value: Ttest_indResult(statistic=0.054352978972766676, pvalue=0.9569116669285772)

cronbach_df_simple.query("agent == 'Claude 2.1' and condition == 'qv'").shape[0] + cronbach_df_simple.query("agent == 'Claude 2.1' and condition == 'vq'").shape[0]

cronbach_df_simple# gpt-4 ttest
ttest_ind(
    cronbach_df_simple.query("agent == 'GPT-4' and condition == 'qv'")["alpha"].values,
    cronbach_df_simple.query("agent == 'GPT-4' and condition == 'vq'")["alpha"].values,
)
# old value: Ttest_indResult(statistic=2.4887215445023743, pvalue=0.016869329019227434)

cronbach_df_simple.query("agent == 'GPT-4' and condition == 'qv'").shape[0] + cronbach_df_simple.query("agent == 'GPT-4' and condition == 'vq'").shape[0]

Calculating Cohen-D for GPT-4

def cohen_d(x, y):
    nx = len(x)
    ny = len(y)
    dof = nx + ny - 2
    return (np.mean(x) - np.mean(y)) / np.sqrt(
        ((nx - 1) * np.std(x, ddof=1) ** 2 + (ny - 1) * np.std(y, ddof=1) ** 2) / dof
    )

cohen_d(
    cronbach_df_simple.query("agent == 'GPT-4' and condition == 'qv'")["alpha"].values,
    cronbach_df_simple.query("agent == 'GPT-4' and condition == 'vq'")["alpha"].values,
)


## ANOVA for agent

cronbach_overall_simple = cronbach_df_simple.query("condition == 'overall'")

formula = "alpha ~ C(agent, Treatment('Human'))"
model = smf.ols(formula, cronbach_overall_simple).fit()
aov_table = sm.stats.anova_lm(model, typ=2)
print(aov_table)

cronbach_overall_simple.groupby(["agent", "scale"])["alpha"].mean()

model.summary2()

## ANOVA AGENT * Instrument

formula_scale = "alpha ~ C(agent, Treatment('Human')) * C(scale)"
model_scale = smf.ols(formula_scale, cronbach_overall_simple).fit()
aov_table_scale = sm.stats.anova_lm(model_scale, typ=2)
round(aov_table_scale, 3)

res = pairwise_tukeyhsd(cronbach_df_simple["alpha"], cronbach_df_simple["agent"])
print(res)

model_scale.summary()

## AGENT * FOUNDATION

formula_found = "alpha ~ C(agent, Treatment('Human')) * C(foundation)"
model_found = smf.ols(formula_found, cronbach_df_simple.query("foundation != 'Liberty'")).fit()

aov_table_found = sm.stats.anova_lm(model_found, typ=2)
round(aov_table_found, 3)

# Regression models

df = pd.read_csv(data_dir / "foundation_cond_agg.csv")

# renaming columns to use same terminology
df.rename(
    columns={
        "MFQ_harm": "MFQ_Care",
        "MFQ_ingroup": "MFQ_Loyalty",
        "MFQ_authority": "MFQ_Authority",
        "MFQ_purity": "MFQ_Purity",
        "MFQ_fairness": "MFQ_Fairness",
    },
    inplace=True,
)

df.head(2)

Applying StandardScaler

scaler = StandardScaler()

# apply on columns starting with MFQ or MFV
df.loc[
    :, df.columns.str.startswith("MFQ") | df.columns.str.startswith("MFV")
] = scaler.fit_transform(
    df.loc[:, df.columns.str.startswith("MFQ") | df.columns.str.startswith("MFV")]
)

# rename columns to standardize under care, loyalty etc

formula = "Q('{}') ~ MFQ_Authority + MFQ_Care + MFQ_Fairness + MFQ_Loyalty + MFQ_Purity"

dependents = [x for x in df.columns if x.startswith("MFV_")]
dependents.sort()

df.columns

## Models for Claude

models = list()

for i in dependents:
    mod = smf.ols(
        formula=formula.format(i),
        data=df.query("agent == 'Claude 2.1'"),
    )
    res = mod.fit()
    models.append(res)

output = summary_col(models, stars=True,)

output

## GPT-4

models = list()

for i in dependents:
    mod = smf.ols(
        formula=formula.format(i),
        data=df.query("agent == 'GPT-4' and condition == 'vq'"),
    )
    res = mod.fit()
    models.append(res)

output = summary_col(models, stars=True)

output

models = list()

for i in dependents:
    mod = smf.ols(
        formula=formula.format(i),
        data=df.query("agent == 'GPT-4' and condition == 'qv'"),
    )
    res = mod.fit()
    models.append(res)

output_qv = summary_col(models, stars=True)

output_qv

models = list()

for i in dependents:
    mod = smf.ols(
        formula=formula.format(i),
        data=df.query("agent == 'GPT-4' and condition == 'vq'"),
    )
    res = mod.fit()
    models.append(res)

output_vq = summary_col(models, stars=True)

output_vq

# Statistics - Including MFQ Divisions

## Condition T-test

cronbach_df_parts = cronbach_df.query("scale != 'MFQ'")

# claude ttest
ttest_ind(
    cronbach_df_parts.query("agent == 'Claude 2.1' and condition == 'qv'")["alpha"].values,
    cronbach_df_parts.query("agent == 'Claude 2.1' and condition == 'vq'")["alpha"].values,
)

# old value: Ttest_indResult(statistic=0.054352978972766676, pvalue=0.9569116669285772)

# gpt-4 ttest
ttest_ind(
    cronbach_df_parts.query("agent == 'GPT-4' and condition == 'qv'")["alpha"].values,
    cronbach_df_parts.query("agent == 'GPT-4' and condition == 'vq'")["alpha"].values,
)
# old value: Ttest_indResult(statistic=2.4887215445023743, pvalue=0.016869329019227434)

Calculating Cohen-D for GPT-4

def cohen_d(x, y):
    nx = len(x)
    ny = len(y)
    dof = nx + ny - 2
    return (np.mean(x) - np.mean(y)) / np.sqrt(
        ((nx - 1) * np.std(x, ddof=1) ** 2 + (ny - 1) * np.std(y, ddof=1) ** 2) / dof
    )

cohen_d(
    cronbach_df_parts.query("agent == 'GPT-4' and condition == 'qv'")["alpha"].values,
    cronbach_df_parts.query("agent == 'GPT-4' and condition == 'vq'")["alpha"].values,
)

# previous 0.7503777791625594

## ANOVA for agent

formula = "alpha ~ C(agent, Treatment('Human'))"
model = smf.ols(formula, cronbach_df_parts).fit()
aov_table = sm.stats.anova_lm(model, typ=2)
print(aov_table)

cronbach_df_parts.groupby("agent")["alpha"].mean()

model.summary2()

formula_scale = "alpha ~ C(agent, Treatment('Human')) * C(scale)"
model_scale = smf.ols(formula_scale, cronbach_df_parts).fit()
aov_table_scale = sm.stats.anova_lm(model_scale, typ=2)
aov_table_scale

formula_found = "alpha ~ C(agent, Treatment('Human')) * C(foundation)"
model_found = smf.ols(formula_found, cronbach_df_parts).fit()
aov_table_found = sm.stats.anova_lm(model_found, typ=2)
aov_table_found

t = pairwise_tukeyhsd(cronbach_df_parts["alpha"], cronbach_df_parts["agent"])
print(t)

When adjusting for multiple comparisons, the effect of agent is not significant. Furthermore, LLMs mean was slightliy higher than that of humans.

Only Part 1

cronbach_mfv_mfq1 = cronbach_df_parts.query("scale !=  'MFQ - Part 2'")
cronbach_mfv_mfq2 = cronbach_df_parts.query("scale !=  'MFV - Part 1'")

formula_scale = "alpha ~ C(agent, Treatment('Human')) * C(scale)"
model_scale = smf.ols(formula_scale, cronbach_mfv_mfq1).fit()
aov_table_scale = sm.stats.anova_lm(model_scale, typ=2)
aov_table_scale

print(pairwise_tukeyhsd(cronbach_mfv_mfq1["alpha"], cronbach_mfv_mfq1["agent"]))

formula_scale = "alpha ~ C(agent, Treatment('Human')) * C(scale)"
model_scale = smf.ols(formula_scale, cronbach_mfv_mfq2).fit()
aov_table_scale = sm.stats.anova_lm(model_scale, typ=2)
aov_table_scale
